chore: moved markets_fold carousel to features folder
