+++ crowdin/messages.json
@@ -1570,6 +1570,12 @@
   "-1197864059": "Create free demo account",
   "-795459217": "Go to live demo",
   "-2145244263": "This field is required",
+  "-1360155081": "Take part in the world’s largest financial market where more than $5 trillion worth of currencies are bought and sold each day.",
+  "-1235219375": "Trade share price movements of big brands and predict broader market trends with indices that measure the overall performance of a market.",
+  "-1959484303": "Cryptocurrencies",
+  "-295883723": "Trade on the rising and falling prices of the most popular cryptocurrencies without the need to own a digital wallet.",
+  "-305466448": "Trade your favourite currency against a basket of major currencies and benefit from reduced risk and volatility.",
+  "-362324454": "Commodities",
   "-319687255": "Deriv GO",
   "-1879666853": "Deriv MT5",
   "-1344568737": "The all-in-one CFD trading platform.",
@@ -1585,12 +1591,6 @@
   "-64934031": "The DMT5 Financial account offers you leverage to trade contracts for difference (CFDs) on forex, stocks, stock indices, commodities, synthetic indices, and cryptocurrencies.",
   "-240686434": "Reliable.",
   "-396178446": "100+ tradeable assets",
-  "-1360155081": "Take part in the world’s largest financial market where more than $5 trillion worth of currencies are bought and sold each day.",
-  "-1235219375": "Trade share price movements of big brands and predict broader market trends with indices that measure the overall performance of a market.",
-  "-1959484303": "Cryptocurrencies",
-  "-295883723": "Trade on the rising and falling prices of the most popular cryptocurrencies without the need to own a digital wallet.",
-  "-305466448": "Trade your favourite currency against a basket of major currencies and benefit from reduced risk and volatility.",
-  "-362324454": "Commodities",
   "-1807088090": "Settle all contracts fairly, accurately, and promptly",
   "-666356424": "We will design our cashier systems to work as seamlessly as possible, processing deposits and withdrawals efficiently and promptly. As far as possible, we will minimise delays imposed by any AML or anti-fraud requirements.",
   "-1096649509": "Provide reliable online trading with high uptime, good security, and low latency",
