+++ src/features/home/carousels/markets_folds/types.ts
